# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API MOBILE",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/payment/verify-apple-receipt": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Verify Apple receipt and save course payment",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Receipt data from Apple purchase",
                        "required": true,
                        "schema": {
                            "properties": {
                                "receipt_data": {
                                    "type": "string",
                                    "example": "MIIUKQYJKoZIhvcNAQcCoIIUGjCCFBYCAQExCzAJBgUrDgMCGgUA..."
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment verified successfully",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Payment verified and saved successfully!"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid receipt data",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "error": {
                                    "type": "string",
                                    "example": "Invalid receipt data"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "409": {
                        "description": "Bad Request - Payments already exists",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "error": {
                                    "type": "string",
                                    "example": "Payment already exists"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "error": {
                                    "type": "string",
                                    "example": "Error occurred while verifying receipt"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearer_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}